Extract desktop relay service worker helpers
